apm: Refactor service impl. to support multiple ports.
